auth: Update provider to add device token auth (#10)
On first run:

```
│ Error: NPS Provider Authentication error
│
│   with provider["registry.terraform.io/northpolesec/nps"],
│   on main.tf line 12, in provider "nps":
│   12: provider "nps" {
│
│ Not logged in. Run `/Users/rah/go/bin/terraform-provider-nps -login` to login
```

When running the provider:

```
Attempting to automatically open the SSO authorization page in your default browser.
If the browser does not open or you wish to use a different device to authorize this request, open the following URL:

https://seamless-valley-15-staging.authkit.app/device?user_code=MWVH-FNRR

Waiting for token...
```

Note: the web browser has opened in the background to confirm the
access. Also, the URL will be signin.workshop.cloud in prod.

After confirming in the web UI:

```
Waiting for token...
Successfully logged in
```

After this, Terraform runs normally and any actions taken are associated
with the logged in user (e.g. rah@northpole.security).

There are also some unrelated changes to fix the Rule and APIKey
resource types to match changes in Workshop, along with validation of
fields because I kept running into RPC errors while testing.